jecrc_chatbot/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-groq
faiss-cpu
sentence-transformers
pypdf
beautifulsoup4